Added addition of points in percents

+++ BotAnbotip/Bot/Handlers/MessageHandler.cs
@@ -54,7 +54,7 @@ namespace BotAnbotip.Bot.Handlers
             {
                 if (!DataManager.UserProfiles.Value.ContainsKey(message.Author.Id))
                     DataManager.UserProfiles.Value.Add(message.Author.Id, new UserProfile(message.Author.Id));
-                await DataManager.UserProfiles.Value[message.Author.Id].AddPoints((int)ActionsCost.Message);
+                await DataManager.UserProfiles.Value[message.Author.Id].AddPoints((long)ActionsCost.Message);
                 await DataManager.UserProfiles.SaveAsync();
             });
         }
